+++ src/ifba/ads/model/AppTeste.java
@@ -10,12 +10,29 @@ public class AppTeste {
 		configuracao.adicionarEquipamentosAUnidade(EquipamentoEnum.CAMERADEVIDEO);
 		configuracao.adicionarEquipamentosAUnidade(EquipamentoEnum.MEDIDORDEC02);
 		
-		UnidadeMovel unidade = new UnidadeEuclidiana("teste", 1234,4321, configuracao);
-		System.out.println(unidade);
+		ConfiguracaoDaUnidade configuracao2 = new ConfiguracaoDaUnidade();
 		
-		SistemaDeMonitoramento monitoramento = new SistemaDeMonitoramento();
+		configuracao2.adicionarEquipamentosAUnidade(EquipamentoEnum.MEDIDORDEC02);
 		
+		ConfiguracaoDaUnidade configuracao3 = new ConfiguracaoDaUnidade();
+		configuracao3.adicionarEquipamentosAUnidade(EquipamentoEnum.CAMERADEVIDEO);
+		
+		
+		UnidadeMovel unidade = new UnidadeEuclidiana("teste", 4,2, configuracao);
+		UnidadeMovel unidade3 = new UnidadeEuclidiana("teste3", 4,2, configuracao3);		
+		UnidadeMovel unidadeM = new UnidadeManhattan("teste2", 42,24, configuracao2);
 
+		AreaMonitorada areaMonitorada = new AreaMonitorada();
+		
+		areaMonitorada.inserirUnidades(unidade);
+		areaMonitorada.inserirUnidades(unidadeM);
+		areaMonitorada.inserirUnidades(unidade3);
+		
+		
+		System.out.println(areaMonitorada.monitorar(7, 5, false, false, false, false));
+
+		
 	}
 
+	
 }