+++ src/ifba/ads/model/ComandarMonitoramento.java
@@ -4,6 +4,7 @@ public interface ComandarMonitoramento {
 
 	public String monitorar(float latitude, float longitude, boolean cameraDeVideo, boolean termometro, boolean MedidordeCo2,
 			boolean medidorDeMetano);
-	public void atualizarLocalizacao(float latitude, float longitude);
+	public void atualizarLocalizacao(UnidadeMovel unidade, float latitude, float longitude);
+	public void inserirUnidades(UnidadeMovel unidade);
 	
 }