+++ src/ifba/ads/model/UnidadeEuclidiana.java
@@ -2,14 +2,18 @@ package ifba.ads.model;
 
 public class UnidadeEuclidiana extends UnidadeMovel{
 
-	public UnidadeEuclidiana(String id, long latitude, long longitude, ConfiguracaoDaUnidade configuracao) {
+	
+	public UnidadeEuclidiana(String id, float latitude, float longitude, ConfiguracaoDaUnidade configuracao) {
 		super(id, latitude, longitude, configuracao);
 	}
 
 	@Override
-	public long distanciaEntreOrigemEDestino() {
-
-		return 0;
+	public float distanciaEntreOrigemEDestino(float latitudeDestino, float longitudeDestino) {
+		long obterValorQuadradoLatitude = (long) Math.pow((latitudeDestino - this.getLatitude()),2);
+		long obterValorQuadradoLongitude = (long) Math.pow((longitudeDestino - this.getLongitude()),2);
+		
+		return (float) Math.sqrt(obterValorQuadradoLatitude + obterValorQuadradoLongitude);
+	
 	}
 
 }