+++ Release Notes.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## 0.0.3
++ Implementação da área monitorada
+
+## 0.0.2
++ Implementação do calculo da distancia
++ Implementação do metodo para atualizar a localização
+
 ## 0.0.1
 + Estrutura da classe de negócio 
 + Implementação dos tipos de unidade